Move View to own namespace
