+++ src/components/feature/features.js
@@ -1,33 +1,9 @@
+import View from "./view.js";
+
 const fetchOptions = {
     mode: "cors",
     errorMessage: "ATENÇÃO: A funcionalidade selecionada não está disponível"
 };
-
-function View(context, hiccup) {
-    const dom = context.Dom(context, hiccup),
-          node = context.document.getElementById(context.renderNodes.featuresBody),
-          rendered = dom.appendToDomNode(node),
-          state = { visible: true, };
-
-    function show() {
-        if ( ! state.visible ) {
-            node.appendChild(rendered);
-            state.visible = true;
-        }
-    }
-
-    function hide() {
-        if ( state.visible ) {
-            node.removeChild(rendered);
-            state.visible = false;
-        }
-    }
-
-    return {
-        show,
-        hide,
-    };
-}
 
 async function Feature(context, featureId) {
     const { httpClient, config, api, Dom } = context,
